app/requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
CacheControl==0.14.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels-auth-token-middlewares==1.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==5.1.2
django-allauth==0.57.0
django-appconf==1.0.6
django-channels-jwt==0.0.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
execnet==2.1.1
executing==2.1.0
fcm-django==2.2.1
firebase-admin==6.6.0
flower==2.0.1
geomet==0.2.1.post1
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
ipython==8.29.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
matplotlib-inline==0.1.7
model-bakery==1.20.0
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-django==4.9.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.1
rsa==4.9
scylla-driver==3.28.0
service-identity==24.2.0
setuptools==75.3.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.3
swapper==1.4.0
tornado==6.4.1
traitlets==5.14.3
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.1.1
